--- lib/core/models.dart
@@ -1,1 +1,0 @@
-// models
